requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0

# Database
sqlalchemy>=2.0
aiosqlite>=0.19

# HTTP/API
httpx>=0.25
aiohttp>=3.9

# Amazon
paapi5-python-sdk>=1.0  # Amazon Product Advertising API
beautifulsoup4>=4.12    # Fallback scraping
lxml>=5.0

# Pinterest
pinterest-api>=0.1

# Image Processing
Pillow>=10.0
requests>=2.31

# Scheduling
apscheduler>=3.10

# Dashboard
streamlit>=1.30
plotly>=5.18
pandas>=2.0

# AI/LLM
anthropic>=0.18

# Utilities
rich>=13.0              # Pretty terminal output
loguru>=0.7             # Logging
tenacity>=8.2           # Retry logic
